refactor stripe
