+++ pages/api/products.ts
@@ -1,11 +1,11 @@
 // pages/api/products.ts
 import Stripe from 'stripe';
+import { stripe } from '@/lib/stripe';
 import { NextApiRequest, NextApiResponse } from 'next';
 
 if(!process.env.STRIPE_SECRET_KEY) {
   throw new Error("STRIPE_SECRET_KEY is not defined")
 }
-const stripe = new Stripe(process.env.STRIPE_SECRET_KEY, {apiVersion: "2025-03-31.basil"})
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   if (req.method !== 'GET') {
